add Pime number check and Fibbonci feature
