+++ homework/Algorithms/Lesson1/Lesson1.sln
@@ -5,10 +5,6 @@ VisualStudioVersion = 16.0.32126.315
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lesson1", "Lesson1\Lesson1.csproj", "{E43415F7-C728-4AB8-A8D7-840502C1D0EE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lesson1.1", "Lesson1.1\Lesson1.1.csproj", "{0BE9F84E-07DD-4F16-B18F-A5F92B4CACBC}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lesson1.2", "Lesson1.2\Lesson1.2.csproj", "{92916223-BCBD-4A44-847F-C8A65C57A136}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -19,14 +15,6 @@ Global
 		{E43415F7-C728-4AB8-A8D7-840502C1D0EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E43415F7-C728-4AB8-A8D7-840502C1D0EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E43415F7-C728-4AB8-A8D7-840502C1D0EE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0BE9F84E-07DD-4F16-B18F-A5F92B4CACBC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0BE9F84E-07DD-4F16-B18F-A5F92B4CACBC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0BE9F84E-07DD-4F16-B18F-A5F92B4CACBC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0BE9F84E-07DD-4F16-B18F-A5F92B4CACBC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{92916223-BCBD-4A44-847F-C8A65C57A136}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{92916223-BCBD-4A44-847F-C8A65C57A136}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{92916223-BCBD-4A44-847F-C8A65C57A136}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{92916223-BCBD-4A44-847F-C8A65C57A136}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
