+++ homework/Algorithms/Lesson1/Lesson1/FibCase.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Lesson1
+{
+    public class FibCase
+    {
+        public int Input { get; set; }
+        public int ExpectedResult { get; set; }
+        public string ExpectedException { get; set; }
+    }
+}
